pequeños errores

+++ script/compose.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 set -o pipefail
 ##se instalara complemento docker "docker compose"
-sudo curl -L "https://github.com/docker/compose/releases/download/1.24.0/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sudo chmod +x /usr/local/bin/docker-compose
-sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
+if [ -e /usr/local/bin/docker-compose ]; then
+    echo "ya se encuentra docker-compose instalado"
+else
+    sudo curl -L "https://github.com/docker/compose/releases/download/1.24.0/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+    sudo chmod +x /usr/local/bin/docker-compose
+    sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
+fi
